Added ability to create new essays in the side menu

+++ app/src/utilities/file-crawler.js
@@ -36,7 +36,7 @@ class FileCrawler {
   }
 
   removeFile(fileName) {
-    fs.removeSync(path.join(this.dir, fileName));
+    if (fileName !== undefined) fs.removeSync(path.join(this.dir, fileName));
   }
 
   writeFile(fileName, data) {
